+++ switch.py
@@ -86,6 +86,22 @@ class Switch:
         self.draw2 = False
         self.draw4 = False
 
+    def special_event(self, player):
+        if self.skip:
+            # return without performing any discard
+            self.skip == False
+            UI.print_message('{} is skipped.'.format(player.name))
+        elif self.draw2:
+            # draw two cards
+            picked = self.pick_up_card(player, 2)
+            self.draw2 == False
+            UI.print_message('{} draws {} cards.'.format(player.name, picked))
+        elif self.draw4:
+            # draw four cards
+            picked = self.pick_up_card(player, 4)
+            self.draw4 == False
+            UI.print_message('{} draws {} cards.'.format(player.name, picked))
+
     def run_player(self, player):
         """Process a single player's turn.
 
@@ -103,20 +119,8 @@ class Switch:
         called to draw from stock.
         """
         # apply any pending penalties (skip, draw2, draw4)
-        if self.skip:
-            # return without performing any discard
-            self.skip = False
-            UI.print_message('{} is skipped.'.format(player.name))
-        elif self.draw2:
-            # draw two cards
-            picked = self.pick_up_card(player, 2)
-            self.draw2 = False
-            UI.print_message('{} draws {} cards.'.format(player.name, picked))
-        elif self.draw4:
-            # draw four cards
-            picked = self.pick_up_card(player, 4)
-            self.draw4 = False
-            UI.print_message('{} draws {} cards.'.format(player.name, picked))
+
+        self.special_event(player)
 
         top_card = self.discards[-1]
         hand_sizes = [len(p.hand) for p in self.players]